fix: fix dual screen
