+++ packages/shared/src/modules/DualScreenInfo/useDualScreenInfo.ts
@@ -1,48 +1,4 @@
-/**
- * React hook for dual screen info
- */
-
-import { useCallback, useEffect, useMemo, useState } from 'react';
-
-import platformEnv from '../../platformEnv';
-
-import {
-  addSpanningListener,
-  isDualScreenDevice as checkIsDualScreenDevice,
-  isSpanning as checkIsSpanning,
-  getHingeBounds as fetchHingeBounds,
-  getWindowRects as fetchWindowRects,
-} from './DualScreenInfo';
-
-import type { IDualScreenInfoRect, ISpanningEvent } from './DualScreenInfo';
-
-export interface IUseDualScreenInfoResult {
-  isDualScreenDevice: boolean;
-  isSpanning: boolean;
-  windowRects: IDualScreenInfoRect[];
-  hingeBounds: IDualScreenInfoRect | null;
-  isLoading: boolean;
-  refresh: () => Promise<void>;
-}
-
-/**
- * Hook to use dual screen information in React components
- * Automatically subscribes to spanning changes and provides current state
- *
- * @example
- * ```tsx
- * function MyComponent() {
- *   const { isDualScreenDevice, isSpanning, windowRects } = useDualScreenInfo();
- *
- *   return (
- *     <View>
- *       <Text>Dual Screen: {isDualScreenDevice ? 'Yes' : 'No'}</Text>
- *       <Text>Spanning: {isSpanning ? 'Yes' : 'No'}</Text>
- *     </View>
- *   );
- * }
- * ```
- */
+import type { IUseDualScreenInfoResult } from './type';
 
 const DEFAULT_RESULT: IUseDualScreenInfoResult = {
   isDualScreenDevice: false,
@@ -53,77 +9,4 @@ const DEFAULT_RESULT: IUseDualScreenInfoResult = {
   refresh: async () => {},
 };
 
-let IS_DUAL_SCREEN_DEVICE = false;
-export const useDualScreenInfo = platformEnv.isNativeAndroid
-  ? (): IUseDualScreenInfoResult => {
-      const [isDualScreenDevice, setIsDualScreenDevice] = useState(
-        IS_DUAL_SCREEN_DEVICE,
-      );
-      const [isSpanning, setIsSpanning] = useState(false);
-      const [windowRects, setWindowRects] = useState<IDualScreenInfoRect[]>([]);
-      const [hingeBounds, setHingeBounds] =
-        useState<IDualScreenInfoRect | null>(null);
-      const [isLoading, setIsLoading] = useState(true);
-
-      const refresh = useCallback(async () => {
-        try {
-          const [isDualScreen, spanning, rects, hinge] = await Promise.all([
-            checkIsDualScreenDevice(),
-            checkIsSpanning(),
-            fetchWindowRects(),
-            fetchHingeBounds(),
-          ]);
-
-          setIsDualScreenDevice(isDualScreen);
-          IS_DUAL_SCREEN_DEVICE = isDualScreen;
-          setIsSpanning(spanning);
-          setWindowRects(rects);
-          setHingeBounds(hinge);
-        } catch (error) {
-          console.warn('useDualScreenInfo refresh error:', error);
-        } finally {
-          setIsLoading(false);
-        }
-      }, []);
-
-      useEffect(() => {
-        // Initial load
-        void refresh();
-
-        // Subscribe to spanning changes
-        const subscription = addSpanningListener((event: ISpanningEvent) => {
-          setIsSpanning(event.isSpanning);
-          // Refresh window rects and hinge bounds when spanning changes
-          void Promise.all([fetchWindowRects(), fetchHingeBounds()]).then(
-            ([rects, hinge]) => {
-              setWindowRects(rects);
-              setHingeBounds(hinge);
-            },
-          );
-        });
-
-        return () => {
-          subscription.remove();
-        };
-      }, [refresh]);
-
-      return useMemo(
-        () => ({
-          isDualScreenDevice,
-          isSpanning,
-          windowRects,
-          hingeBounds,
-          isLoading,
-          refresh,
-        }),
-        [
-          isDualScreenDevice,
-          isSpanning,
-          windowRects,
-          hingeBounds,
-          isLoading,
-          refresh,
-        ],
-      );
-    }
-  : () => DEFAULT_RESULT;
+export const useDualScreenInfo = () => DEFAULT_RESULT;
